requirements.txt
 asgiref-3.8.1
 django-5.1.2 
 sqlparse-0.5.1 
 tzdata-2024.2
 Pillow-11.0.0